Mores test added and code splitting in VideoDetail
